add some responsiveness to navbar and hero
