+++ src/components/Navigation/NavigationLinks.jsx
@@ -1,27 +1,9 @@
 import React from 'react';
-import logo from '../../images/logo.png';
 
 const Navigation = () => {
-
-  const toTop = () => {
-    window.scroll({
-      top: 0,
-      left: 0,
-      behavior: 'smooth',
-    });
-  };
   
   return (
-    <div>
-      <div className="p-3 flex justify-center">
-        <button onClick={() => toTop()}>
-          <img
-            className="cursor-pointer "
-            src={logo}
-            alt="Pilgrim's Finest Logo"
-          ></img>
-        </button>
-      </div>
+    <div className="hidden md:flex justify-around col-span-3">
       <div className="p-3 flex justify-center">
         <span className="hover:text-yellow-500 cursor-pointer">Our Products</span>
       </div>
